auto detect camera sensor existing in unittest; simplify task's codes.
