+++ python/social_bot/teacher.py
@@ -56,8 +56,19 @@ class Task(object):
     """
 
     def __init__(self):
+        self._agent = None
+        self._world = None
+        self._agent_name = None
         self.reward_weight = 1.0
 
+    def setup(self, world, agent_name):
+        """
+        Setting things up during the initialization
+        """
+        self._world = world
+        self._agent = self._world.get_agent()
+        self._agent_name = agent_name
+        
     @abstractmethod
     def run(self):
         """
